pyproject.toml
[project]
name = "akashi-api"
version = "0.1.0"
description = "AKASHI MAM - REST API"
authors = [
    {name = "Tiago Cunha", email = "tiago@gruponortecomunicacao.com.br"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",

    # Validation & Settings
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",

    # Object Storage (S3/MinIO)
    "boto3>=1.34.0",
    "aioboto3>=12.0.0",

    # Queue
    "celery>=5.3.6",
    "redis>=5.0.1",

    # Media Processing
    "ffmpeg-python>=0.2.0",

    # Utilities
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",
    "ruff>=0.1.11",
    "black>=24.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "ipython>=8.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require running API)",
]
